Fix Payments

+++ SWP391_G5_EventFlowerExchange_BE/src/main/java/com/SWP391_G5_EventFlowerExchange/LoginAPI/entity/Delivery.java
@@ -18,7 +18,7 @@ public class Delivery {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long deliveryID;
+    private int deliveryID;
 
     @Column(nullable = false)
     private LocalDateTime deliveryDate;
